requirements.txt
streamlit
streamlit-webrtc
openai
speechrecognition
pydub
pillow
requests
beautifulsoup4
google-search-results
streamlit-camera-input-live
gtts
pygame
streamlit-chat
opencv-python-headless
numpy
audio-recorder-streamlit
streamlit-option-menu
streamlit-lottie
plotly
datetime